requirements.txt
asgiref==3.2.7
click==7.1.2
cycler==0.10.0
Cython==0.29.20
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
gunicorn==20.0.4
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.1
nltk==3.5
numpy==1.18.5
pandas==1.0.4
Pillow==7.1.2
psycopg2==2.8.5
pyparsing==2.4.7
pytesseract==0.3.4
python-datauri==0.2.8
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
six==1.15.0
sqlparse==0.3.1
tqdm==4.46.1
whitenoise==5.1.0
wordcloud==1.7.0
